requirements.txt
configparser==5.0.1
pybase62==0.4.3
